Additions to main UI window
Added message box to inform user when they have attempted to use an unavailable feature that is still in development.

Parented the main ui window to the main maya window and changed the closing functionality so when the 'X' button is pressed the window is only hidden. To compeletely shut the application the user must now press 'File->Exit'.