Added documentation to several functions.
Fixed a bug in the outliner when dragging in new items from the addModules window where if they were dropped onto the world space item they would get parented to it.

Potentially from previous commit:
Forgot to add that I've moved the functions that add modules when dragging from the addModules window to the outliner from addModules to moduleOutliner. I'm passing the outliner a reference to the main ui, giving it access to all the different windows (in this case the addModules window) so it can determine which items need to be added